fixed win32 build
git-svn-id: svn://libnui.net/nui/trunk/nui3@5178 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ tutorials/Additive/src/MainWindow.cpp
@@ -140,7 +140,7 @@ float Blur(float x, float mean, float deviation, float variance)
       variance = 0.1;
     }
   }
-  return pow(exp(-(((x - mean) * (x - mean)) / ((2 * variance)))), 1 / (deviation * sqrt(2 * M_PI)));
+  return powf(exp(-(((x - mean) * (x - mean)) / ((2 * variance)))), 1 / (deviation * sqrt(2 * M_PI)));
 }
 
 bool MainWindow::Draw(nuiDrawContext* pContext)
